/ppbox/branches/1.2.0 /src/ppbox/avformat 整理 AvcNaluHelper Codec 等
git-svn-id: http://svn01/repositories/r01/ppbox/branches/1.2.0@6122 b3a1389e-2159-5c43-9038-463bf877d0d9

+++ mkv/ebml/EBML_Archive.h
@@ -114,7 +114,7 @@ namespace ppbox
 
         private:
             std::vector<EBML_ElementHeader *> headers_;
-            std::vector<boost::uint32_t> sizes_;
+            std::vector<boost::uint64_t> sizes_;
         };
 
         class EBML_ElementIArchive
